Add support for AMD Bolton chipset
Add support for AMD's Bolton chipset.  The SPI controller on the bolton chipset
uses the same 3-bit speed settings as yangtze, but is otherwise the same as the
hudson chips.  Note that the Bolton RRG doesn't specify a speed setting for the
bit setting of 0b111, so I'm assuming that it's the same setting as yangtze.

Signed-off-by: Edward O'Callaghan <eocallaghan@alterapraxis.com>